Add class for person avatar
